start

+++ handlers/table_reservations_handlers.py
@@ -1,9 +1,11 @@
 from datetime import datetime, timedelta, date
 import os
 
-from aiogram import Router
+from aiogram import Router, flags
 from aiogram.types import Message, FSInputFile
 from aiogram.fsm.context import FSMContext
+from aiogram.filters import Filter
+
 
 from data import db_funcs_admin_menu
 from states.states import StateTableReservations, StateMenu
@@ -20,10 +22,51 @@ router = Router()
 """
 
 
+class ProfileStateFilter(Filter):
+    async def __call__(self, msg: Message, state: FSMContext) -> bool:
+        current_state = await state.get_state()
+        prompt = msg.text
+        return current_state in [
+            StateTableReservations.add_table_reservations_booking_start_time_time,
+            StateTableReservations.add_table_reservations_table,
+            StateTableReservations.add_table_reservations_number_of_guests,
+            StateTableReservations.add_table_reservations_confirmation_enter_data
+        ] and (prompt == "Назад" or prompt == "Выбрать другое время")
+
+
+@router.message(ProfileStateFilter())
+@flags.chat_action("typing")
+async def back_table_reservations(msg: Message, state: FSMContext) -> None:
+    user_id = msg.from_user.id
+    now_state = await state.get_state()
+    await state.set_state()
+    datas: dict = await state.get_data()
+    if now_state == StateTableReservations.add_table_reservations_booking_start_time_time:
+        await msg.answer(text=text_reservation.add_table_reservations_booking_start_time_date,
+                         reply_markup=kb_table_reservations.date_enter())
+        await state.set_state(StateTableReservations.add_table_reservations_booking_start_time_date)
+    elif now_state == StateTableReservations.add_table_reservations_table:
+        date_datas = datetime.strptime(datas['date'], '%Y-%m-%d')
+        await msg.answer(text=text_reservation.add_table_reservations_booking_start_time_time,
+                         reply_markup=kb_table_reservations.time_enter(day_date=date_datas))
+        await state.set_state(StateTableReservations.add_table_reservations_booking_start_time_time)
+    elif now_state == StateTableReservations.add_table_reservations_number_of_guests:
+        path_plan = await working_images_funcs.processing_images(datas['tables_close'])
+        plan_img = FSInputFile(path=path_plan)
+        await msg.answer_photo(photo=plan_img,
+                               caption=text_reservation.add_table_reservations_table,
+                               reply_markup=kb_table_reservations.choosing_a_free_table(datas['tables_open']))
+        await state.set_state(StateTableReservations.add_table_reservations_table)
+    elif now_state == StateTableReservations.add_table_reservations_confirmation_enter_data:
+        await msg.answer(text=text_reservation.add_table_reservations_number_of_guests,
+                         reply_markup=kb_table_reservations.number_of_seats(datas['table']))
+        await state.set_state(StateTableReservations.add_table_reservations_number_of_guests)
+
 
 
 @router.message(StateTableReservations.main_table_reservations)
-async def admin_table_reservations(msg: Message, state: FSMContext) -> None:
+@flags.chat_action("typing")
+async def main_table_reservations(msg: Message, state: FSMContext) -> None:
     user_id = msg.from_user.id
     user = User.get(User.user_id == user_id)
     if user.phone is None:
@@ -38,6 +81,7 @@ async def admin_table_reservations(msg: Message, state: FSMContext) -> None:
 
 
 @router.message(StateTableReservations.add_table_reservations_phone)
+@flags.chat_action("typing")
 async def add_table_reservations_phone(msg: Message, state: FSMContext) -> None:
     user_id = msg.from_user.id
     prompt = msg.text
@@ -64,6 +108,7 @@ async def add_table_reservations_phone(msg: Message, state: FSMContext) -> None:
 
 
 @router.message(StateTableReservations.add_table_reservations_booking_start_time_date)
+@flags.chat_action("typing")
 async def add_table_reservations_booking_start_time_date(msg: Message, state: FSMContext) -> None:
     user_id = msg.from_user.id
     prompt = msg.text
@@ -98,6 +143,7 @@ async def add_table_reservations_booking_start_time_date(msg: Message, state: FS
 
 
 @router.message(StateTableReservations.add_table_reservations_booking_start_time_time)
+@flags.chat_action("typing")
 async def add_table_reservations_booking_start_time_time(msg: Message, state: FSMContext) -> None:
     user_id = msg.from_user.id
     prompt = msg.text
@@ -113,13 +159,14 @@ async def add_table_reservations_booking_start_time_time(msg: Message, state: FS
         if check_date:
             prompt = easy_funcs.correction_datas(time=prompt)
             datas['booking_start_time'] = datetime.strptime(datas['date'] + ' ' + prompt, '%Y-%m-%d %H:%M')
-            del datas['date']
+
             tables_close = []
             tables_all = []
             tables_reservations = (TableReservationHistory
                                    .select().where(TableReservationHistory.booking_start_time
-                                                   .between(datas['booking_start_time'] + timedelta(minutes=1),
-                                                            datas['booking_start_time'] + timedelta(hours=2,
+                                                   .between(datas['booking_start_time'] - timedelta(hours=1,
+                                                                                                    minutes=59),
+                                                            datas['booking_start_time'] + timedelta(hours=1,
                                                                                                     minutes=59))
                                                    ).order_by(TableReservationHistory.booking_start_time))
             for tab in tables_reservations:
@@ -128,6 +175,7 @@ async def add_table_reservations_booking_start_time_time(msg: Message, state: FS
                 tables_all.append(tab.number_table)
             tables_open = list(set(tables_all) - set(tables_close))
             tables_open.sort()
+            datas['tables_close'] = tables_close
             datas['tables_open'] = tables_open
             await state.update_data(**datas)
             path_plan = await working_images_funcs.processing_images(tables_close)
@@ -136,7 +184,8 @@ async def add_table_reservations_booking_start_time_time(msg: Message, state: FS
                                    caption=text_reservation.add_table_reservations_table,
                                    reply_markup=kb_table_reservations.choosing_a_free_table(tables_open))
             await state.set_state(StateTableReservations.add_table_reservations_table)
-            os.remove(path=path_plan)
+            if tables_close:
+                os.remove(path=path_plan)
         else:
             await msg.answer(text=text_reservation.err_error
                                   + text_reservation.add_table_reservations_booking_start_time_time,
@@ -145,6 +194,7 @@ async def add_table_reservations_booking_start_time_time(msg: Message, state: FS
 
 
 @router.message(StateTableReservations.add_table_reservations_table)
+@flags.chat_action("typing")
 async def add_table_reservations_table(msg: Message, state: FSMContext) -> None:
     user_id = msg.from_user.id
     prompt = msg.text
@@ -171,23 +221,23 @@ async def add_table_reservations_table(msg: Message, state: FSMContext) -> None:
                     datas['table'] = int(key)
             await state.update_data(**datas)
             await msg.answer(text=text_reservation.add_table_reservations_number_of_guests,
-                             reply_markup=kb_user_profile.back_button())
+                             reply_markup=kb_table_reservations.number_of_seats(datas['table']))
             await state.set_state(StateTableReservations.add_table_reservations_number_of_guests)
         elif prompt_int(prompt):
             datas['table'] = int(prompt)
-            del datas['tables_open']
             await state.update_data(**datas)
             await msg.answer(text=text_reservation.add_table_reservations_number_of_guests,
-                             reply_markup=kb_user_profile.back_button())
+                             reply_markup=kb_table_reservations.number_of_seats(datas['table']))
             await state.set_state(StateTableReservations.add_table_reservations_number_of_guests)
         else:
 
             await msg.answer(text=text_reservation.add_table_reservations_err_table,
-                             reply_markup=kb_user_profile.back_button())
+                             reply_markup=kb_table_reservations.choosing_a_free_table(datas['tables_open']))
             await state.set_state(StateTableReservations.add_table_reservations_table)
 
 
 @router.message(StateTableReservations.add_table_reservations_number_of_guests)
+@flags.chat_action("typing")
 async def add_table_reservations_number_of_guests(msg: Message, state: FSMContext) -> None:
     user_id = msg.from_user.id
     prompt = msg.text
@@ -197,9 +247,17 @@ async def add_table_reservations_number_of_guests(msg: Message, state: FSMContex
         await state.clear()
         await state.set_state(StateMenu.main_menu)
     else:
-        check_date = easy_funcs.checking_data_expression(number_of_guests=prompt)
-        if check_date:
-            datas = await state.get_data()
+        def prompt_check(x, max_quests):
+            try:
+                int(x)
+                x = 1 <= x <= max_quests
+                return bool(x)
+            except ValueError:
+                return False
+        datas = await state.get_data()
+        table = datas['table']
+        max_quests = data_tables[str(table)]['number_of_seats']
+        if prompt_check(int(prompt), max_quests):
             datas['number_of_guests'] = prompt
             table = Table.select().where(Table.number_table == datas["table"]).get()
             table = table.name_table
@@ -210,15 +268,16 @@ async def add_table_reservations_number_of_guests(msg: Message, state: FSMContex
                 number_of_guests=datas['number_of_guests']),
                 reply_markup=kb_table_reservations.yes_no())
             await state.set_state(StateTableReservations.add_table_reservations_confirmation_enter_data)
-        elif not check_date:
+        else:
             await msg.answer(text=text_reservation.err_error
                                   + text_reservation.add_table_reservations_number_of_guests,
-                             reply_markup=kb_user_profile.back_button())
+                             reply_markup=kb_table_reservations.choosing_a_free_table(datas['tables_open']))
             await state.set_state(StateTableReservations.add_table_reservations_number_of_guests)
 
 
 @router.message(StateTableReservations.add_table_reservations_confirmation_enter_data)
-async def admin_add_table_reservations_confirmation_enter_data(msg: Message, state: FSMContext) -> None:
+@flags.chat_action("typing")
+async def add_table_reservations_confirmation_enter_data(msg: Message, state: FSMContext) -> None:
     user_id = msg.from_user.id
     prompt = msg.text
     if prompt == 'Да':
